+ more experimental classes + info log in SA when solution is found + size limitation to OperationHistory
git-svn-id: svn://svn.code.sf.net/p/jcop/code/trunk@25 ca20af45-4a20-473b-94bf-0ef87cbc3059

+++ jcop/src/cz/cvut/felk/cig/jcop/result/render/CSVRender.java
@@ -126,7 +126,7 @@ public class CSVRender implements Render {
                 stringBufferHumanReadable.append("]");
                 csvWriter.write(stringBuffer.toString());
                 csvWriter.write(stringBufferHumanReadable.toString());
-                csvWriter.write(Integer.toString(resultEntry.getBestConfiguration().getOperationHistory().getCounter()));
+                csvWriter.write(Long.toString(resultEntry.getBestConfiguration().getOperationHistory().getCounter()));
                 csvWriter.write(Double.toString(resultEntry.getBestFitness()));
 
                 for (OperationHistory operationHistory : resultEntry.getBestConfiguration().getOperationHistory().getChronologicalList()) {
